requirements.txt
-i https://pypi.org/simple
blinker==1.4
bloop==3.0.0
boto3==1.14.18
botocore==1.17.18
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dateparser==0.7.6
docutils==0.15.2
dynamodb-json==1.3
ecdsa==0.15
humanize==2.5.0
idna==2.10
jmespath==0.10.0
maya==0.6.1
pendulum==2.1.0
pyasn1==0.4.8
python-dateutil==2.8.1
python-jose==3.1.0
pytz==2020.1
pytzdata==2019.3
regex==2020.6.8
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
simplejson==3.17.0
six==1.15.0
snaptime==0.2.4
tzlocal==2.1
urllib3==1.25.9 ; python_version != '3.4'
werkzeug==1.0.1
